% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_controls_json}
\alias{get_controls_json}
\title{Get the options JSON}
\usage{
get_controls_json(type, iso3, options, fallback_values)
}
\arguments{
\item{type}{The type of options to get, model or calibration}

\item{iso3}{The iso3 of the country to get options for}

\item{options}{A named list of options to override those in JSON
template. Must be in the format of a list of control options i.e.
list(
control_id = list(
list(
id = "1",
label = "one"
),
list(
id = "2",
label = "two"
)
),
...
)}

\item{fallback_values}{A list of names values to use as fallbacks if
hardcoded defaults are invalid.}
}
\value{
The complete controls JSON
}
\description{
This gets the JSON for the model or calibration options for a particular
country. The control info will be filled in in this order of precedence
\enumerate{
\item hardcoded default values from \code{default_options.csv}
\item \code{fallback_values} arg (if set)
\item Fallback value in \code{controls.R} (if set)
}
}
