% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{translate_labels}
\alias{translate_labels}
\title{Take a vector of labels and translate}
\usage{
translate_labels(labels, lang = traduire::translator()$language())
}
\arguments{
\item{labels}{Vector of labels, the translation keys}

\item{lang}{Language to return output in}
}
\value{
Translated labels for current language
}
\description{
Take a vector of labels and translate
}
