% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_controls_json}
\alias{get_controls_json}
\title{Get the options JSON}
\usage{
get_controls_json(type, iso3, options, override_values, config = list())
}
\arguments{
\item{type}{The type of options to get, model or calibration}

\item{iso3}{The iso3 of the country to get options for}

\item{options}{A named list of options to override those in JSON
template. Must be in the format of a list of control options i.e.
list(
control_id = list(
list(
id = "1",
label = "one"
),
list(
id = "2",
label = "two"
)
),
...
)}

\item{override_values}{A list of named values to use as default values. This
will take precedent over a hardcoded value if both are set.}

\item{config}{Additional configuration options for returned JSON. If
type is 'model' can set boolean 'include_art' and 'include_anc' to
include options for ART and ANC controls. You can also pass
\code{additional_control_groups} which will be appended to the top of the controls
JSON. \code{additional_control_groups} must be a list representing a valid
control group object. No affect for 'calibration' type.}
}
\value{
The complete controls JSON
}
\description{
This gets the JSON for the model or calibration options for a particular
country. The control info will be filled in in this order of precedence
\enumerate{
\item \code{override_values} arg (if set)
\item hardcoded default values from \code{default_options.csv}
}
}
